requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
joblib==1.4.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.2.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
uritemplate==4.1.1
